Drop the `s` again
I decided to drop the `s` parameter. It currently only ties variables
and references to a `Store`. However, that is marginal benefit.

The original intent of the parameter to tie it to a transaction is of
much higher value. Unfortunately, the transaction-scoped values have the
downside that it's difficult to created "reatained" variables in an
ergonomic way, where you can retain a variable and then rescope the
content to any arbitrary transaction.

The reason that is difficult is because then we'd need two kinds of
`RetainedVar`s one which adds an extra `s` parameter to its data type
when injected into a transaction, and another one that can work with
plain types like `Int`.

Let's punt on introducing more type safety in favor of better ergonomics
with regards to implementing systems on top of `casper`.